index: Extract `data` module
